requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
idna==3.8
jose==1.0.0
lxml==5.3.0
pandoc==2.4
passlib==1.7.4
plumbum==1.8.3
ply==3.11
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyMySQL==1.1.1
python-docx==1.1.2
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing-extensions==4.12.2
uvicorn==0.30.6
